requirements.txt
aws-cdk-lib==2.82.0
constructs>=10.0.0
cdk-nag>=1.0.0,<=2.0.0